requirements.txt
# Backend API
fastapi
uvicorn[standard]
sqlmodel
python-dotenv

# Autenticação JWT
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1
python-multipart

# HTTP / Utils
requests
aiofiles
pytest
